+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added type hints functionality in action method parameters.
 - Added ActionResultException handling in action methods.
 - Added BadRequestResult and BadRequestResultException.
+- Added UnauthorizedResult and UnauthorizedResultException. 
 - Session cookie is marked as "secure" if request is secure (https).
 - Deprecated AbstractActionResult.